Adding support for high quality yield if pro user.
